+++ dotnet/cs/src/Modulation/Sine.cs
@@ -4,7 +4,7 @@
  * Created Date: 13/09/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 13/09/2023
+ * Last Modified: 04/10/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -120,7 +120,7 @@ namespace AUTD3Sharp.Modulation
         }
 
         public static Fourier operator +(Sine a, Sine b)
-            => new Fourier().AddComponent(a).AddComponent(b);
+            => new Fourier(a).AddComponent(b);
 
         public override ModulationPtr ModulationPtr()
         {